Update composer.json to use correct directory for PSR-0 autoloading
When Composer generated the autoloading files, it picks up on the `autoload.psr-0` key from `composer.json`. The `Auryn` namespace is mapped to the `src/` directory which does not exist.

This commit changes the directory to `lib` so the correct autoloading files can be generated.

I am missing something... this used to work until the recent `composer update`?