Allow params with types and default to be injected correctly.
I believe the correct behaviour when parameter has both a type and a default value (which currently can only be null), is to inject the dependency if the injector has been explicitly told to do so through either alias, delegate or share. Currently if a parameter has both a type and a default of null, autowiring of the constructor cannot be done, and instead a custom function to create the class is required to work around the limitation.This fixes #67.